@@ -1,8 +1,7 @@
 package com.example.Entity;
 
 import lombok.*;
-
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Data
@@ -11,16 +10,12 @@ import javax.persistence.*;
 @Getter
 @Setter
 @Table(name="EMPLOYEE_DB")
-public class Model {
+public class EmployeeDTO {
     @Id
-   // @GeneratedValue(strategy = GenerationType.AUTO)
     private long emp_id;
     private String emp_first_name;
     private String emp_last_name;
     private String department;
     private String joining_date;
     private String work_location;
-    
-    
-    
 }
