@@ -10,6 +10,9 @@ public class EmployeeServiceApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(EmployeeServiceApplication.class, args);
 	}
-	
+
+	public void addEmployee(){
+		L
+	}
 
 }
